Fix moment async module load
