+++ demo/nuxt.config.js
@@ -33,10 +33,7 @@ export default {
   css: ['~/assets/sass/main.sass'],
   buildModules: ['@nuxt/typescript-build'],
   modules: [coreModuleDist],
-  plugins: [
-    { src: '~/plugins/quill', mode: 'client' },
-    { src: '~/plugins/cypress', mode: 'client' }
-  ],
+  plugins: [{ src: '~/plugins/cypress', mode: 'client' }],
   router: {
     middleware: ['auth', 'cwa']
   },
